* move classes to upper level * split Context into SchemaConfig + QueryContext (nothing else was changed in this commit)
